requirements.txt
beautifulsoup4==4.14.3
python-dotenv==1.2.1
graphviz==0.20.3
httpx==0.28.1
matplotlib==3.10.0
networkx==3.4.2
numpy==2.1.3
scipy==1.14.1
shapely==2.0.6
tqdm==4.67.1
z3-solver==4.13.3.0
